requirements.txt
beautifulsoup4==4.8.0
cabocha-python==0.69
certifi==2019.6.16
chardet==3.0.4
Django==2.1.11
django-bootstrap-form==3.4
django-bootstrap4==0.0.8
django-cleanup==4.0.0
django-contact-form==1.7
django-contact-form-bootstrap==0.1a1
django-forms-bootstrap==3.1.0
django-user-sessions==1.6.0
idna==2.8
mecab-python-windows==0.996.3
mysql-connector-python==8.0.16
mysqlclient==1.3.12
Pillow==6.1.0
protobuf==3.8.0
PyMySQL==0.9.3
pypugjs==5.8.1
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
